# internal/docs/swagger.yaml
basePath: /
definitions:
  api.chatStreamRequest:
    properties:
      query:
        example: What are the key ideas in my notes?
        type: string
      top_k:
        example: 5
        type: integer
    type: object
...
  api.createNotebookRequest:
    properties:
      description:
        example: Notes on ML papers
        type: string
      title:
        example: My Research
        type: string
    type: object
  api.updateNotebookRequest:
    properties:
      description:
        example: Updated description
        type: string
      title:
        example: Updated Title
        type: string
    type: object
...
  model.Notebook:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      title:
        type: string
      updated_at:
        type: string
      user_id:
        type: string
    type: object
...
host: localhost:8080
info:
  contact: {}
  description: RAG-backed notebook and conversation management service.
  title: Notebook Service API
  version: "1.0"
paths:
  /notebooks/:
    get:
      description: Returns all notebooks belonging to the authenticated user.
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/model.Notebook'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: List notebooks
      tags:
      - notebooks
    post:
      description: Creates a new notebook for the authenticated user.
      parameters:
      - description: Notebook fields
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.createNotebookRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/model.Notebook'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a notebook
      tags:
      - notebooks
  /notebooks/{notebookID}:
    delete:
      description: Permanently deletes a notebook and all its conversations and sources.
      parameters:
      - description: Notebook UUID
        in: path
        name: notebookID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete a notebook
      tags:
      - notebooks
    get:
      description: Returns a single notebook by ID. The caller must own it.
      parameters:
      - description: Notebook UUID
        in: path
        name: notebookID
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Notebook'
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get a notebook
      tags:
      - notebooks
    patch:
      description: Updates the title and/or description of a notebook.
      parameters:
      - description: Notebook UUID
        in: path
        name: notebookID
        required: true
        type: string
      - description: Fields to update
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.updateNotebookRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/model.Notebook'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "403":
          description: Forbidden
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update a notebook
      tags:
      - notebooks
...
  /notebooks/{notebookID}/conversations/{conversationID}/chat:
    get:
      description: Retrieves relevant source chunks, streams an LLM completion as
        Server-Sent Events, and persists the exchange.
      parameters:
      - description: Notebook UUID
        in: path
        name: notebookID
        required: true
        type: string
      - description: Conversation UUID
        in: path
        name: conversationID
        required: true
        type: string
      - description: Query and retrieval options
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/api.chatStreamRequest'
      produces:
      - text/event-stream
      responses:
        "200":
          description: 'SSE stream of {\"token\":\"...\"} events, terminated by data:
            [DONE]'
          schema:
            type: string
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Stream a chat response (SSE)
      tags:
      - conversations
...
securityDefinitions:
  BearerAuth:
    description: JWT issued by Authentik. Prefix with "Bearer ".
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"